fix key for get prices

+++ src/backend/bigspoon/bigspoon/bg_order/models.py
@@ -241,9 +241,9 @@ class Order(models.Model):
                     prices.update(dict([(sec['itemTitle']+'-'+k, v) for k, v in sec['items'].items()]))
                 else:
                     prices.update(sec['items'].copy())
-            return self.dish.price + float(sum([prices[k]*v for k, v in self.modifier_json.items()]))
+            return float(self.dish.price) + sum([prices[k]*v for k, v in self.modifier_json.items()])
         else:
-            return self.dish.price * self.quantity
+            return float(self.dish.price * self.quantity)
 
     def __unicode__(self):
         return "(%s - %s) %s | %s x %s" % (
